+++ test-utils/src/main/kotlin/net/corda/testing/driver/Driver.kt
@@ -286,6 +286,7 @@ fun <DI : DriverDSLExposedInterface, D : DriverDSLInternalInterface, A> genericD
         coerce: (D) -> DI,
         dsl: DI.() -> A
 ): A {
+    initialiseFlowStackSnapshotFactory()
     if (initialiseSerialization) initialiseTestSerialization()
     val shutdownHook = addShutdownHook(driverDsl::shutdown)
     try {
